feat: udapte button with gesture for icon
